LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/access-from-coop-page-to-other_coop-ro.https-expected.txt


FAIL access-from-coop-page-to-other (COOP-RO) assert_equals: expected (string) "coop" but got (undefined) undefined
